+++ framy_generator/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.1.10]
 
+* Added function callbacks! #107
 * Changed showing settings dialog to work for showcase app #116
 * Added hiding navigation drawer on big screens #113
 * Added option to disable Device Preview #128