+++ framy_generator/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.1.10]
+## [0.2.0]
 
 * Added function callbacks! #107
 * Changed showing settings dialog to work for showcase app #116